% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccpSubOptJPEarlyStop.R
\name{ccpSubOptJPEarlyStop}
\alias{ccpSubOptJPEarlyStop}
\title{CCP Sub-Optimization for Joint Penalty with Early Stop}
\usage{
ccpSubOptJPEarlyStop(v0, Qo1, Qo2, Qc1, Qc2, Vo, tau)
}
\arguments{
\item{v0}{A numeric vector; the initial direction.}

\item{Qo1}{A matrix representing the first quadratic form in the objective function.}

\item{Qo2}{A matrix representing the second quadratic form in the objective function.}

\item{Qc1}{A list of matrices representing the first quadratic form in the constraints.}

\item{Qc2}{A list of matrices representing the second quadratic form in the constraints.}

\item{Vo}{A matrix representing orthogonal constraints.}

\item{tau}{A numeric value representing the multiplier of slack variables.}
}
\value{
A list containing:
  \item{v}{The optimized direction vector.}
  \item{slack}{The optimized slack variable values.}
  \item{objval}{The final objective value of the optimization problem.}
}
\description{
This function performs a sub-optimization using CVXR for penalty-based joint optimization.
It formulates and solves a convex optimization problem with quadratic objectives and constraints.
}
