% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJIVEJointStrucEstimateJP.R
\name{DJIVEJointStrucEstimateJP}
\alias{DJIVEJointStrucEstimateJP}
\title{Estimate Joint and Partially Joint Structures using DJIVE Method}
\usage{
DJIVEJointStrucEstimateJP(
  VBars,
  UBars,
  phiBars,
  psiBars,
  rBars,
  dataname,
  theta0 = 45,
  optArgin = list(),
  iprint = 0,
  figdir = ""
)
}
\arguments{
\item{VBars}{A list of adjusted signal row spaces.}

\item{UBars}{A list of orthogonal signal column spaces (optional, can be NULL).}

\item{phiBars}{A numeric vector of perturbation angles for each data matrix.}

\item{psiBars}{A numeric vector of additional perturbation angles for column spaces.}

\item{rBars}{A numeric vector of adjusted signal ranks for each data matrix.}

\item{dataname}{A character vector of data matrix names.}

\item{theta0}{A numeric value representing the base angle for perturbation. Default is 45 degrees.}

\item{optArgin}{A list of optional arguments for optimization.}

\item{iprint}{An integer for verbosity level. Default is 0.}

\item{figdir}{A character string indicating where to save output figures.}
}
\value{
A list containing:
  \item{outMap}{A mapping between joint block index set and estimated partially shared structures.}
  \item{keyIdxMap}{A mapping between joint block indices and data blocks.}
  \item{anglesMap}{A mapping between joint block indices and estimated angles.}
  \item{jointBlockOrder}{A record of the order of the joint blocks.}
}
\description{
The function `DJIVEJointStrucEstimateJP` establishes a DC programming problem to estimate joint
and partially joint structures in a set of adjusted signal row spaces using the Penalty CCP algorithm.
}
