% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentileMarcenkoPastur.R, R/ksOpt.R
\name{incMarPas}
\alias{incMarPas}
\title{Incremental Marcenko-Pastur Function}
\usage{
incMarPas(x0, beta, gamma)

incMarPas(x0, beta, gamma)
}
\arguments{
\item{x0}{A numeric value representing the lower limit of integration.}

\item{beta}{A numeric value indicating the ratio of columns to rows in the data matrix.}

\item{gamma}{A numeric value specifying the power to which the function should be raised during integration.}
}
\value{
The value of the integral, representing the cumulative probability up to `x0`.

A numeric value of the integrated Marčenko-Pastur function.
}
\description{
This function computes the integral of the Marcenko-Pastur density function from `x0` to the upper bound of the spectrum, which is used to find the cumulative probability up to `x0`.

Computes the incomplete Marčenko-Pastur distribution function, which is used in estimating noise levels.
}
