% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltyCCPJEarlyStop.R
\name{penaltyCCPJPEarlyStop}
\alias{penaltyCCPJPEarlyStop}
\title{Penalty CCP Joint Penalty with Early Stop Optimization}
\usage{
penaltyCCPJPEarlyStop(v0, Qo1, Qo2, Qc1, Qc2, Vorth, optArgin = list())
}
\arguments{
\item{v0}{A numeric vector; the initial direction.}

\item{Qo1}{A matrix representing the first quadratic form in the objective function.}

\item{Qo2}{A matrix representing the second quadratic form in the objective function.}

\item{Qc1}{A list of matrices representing the first quadratic form in the constraints.}

\item{Qc2}{A list of matrices representing the second quadratic form in the constraints.}

\item{Vorth}{A matrix of orthogonal constraints.}

\item{optArgin}{A list of optimization tuning parameters: \code{tau0}, \code{tau_max},
\code{mu}, \code{t_max}, \code{tol}, \code{delta}. Default values are used if not provided.}
}
\value{
A list with the following components:
  \item{opt_v}{The optimized direction vector.}
  \item{cache_v}{Cache of all vectors from each iteration.}
  \item{cache_cvx_objval}{Cache of objective values from each iteration.}
  \item{cache_slack}{Cache of slack variable values from each iteration.}
  \item{converge}{Flag indicating if the optimization converged.}
}
\description{
This function performs optimization using a penalty-based CCP (Convex Concave Procedure)
to estimate joint structures in high-dimensional data with an early stop mechanism.
}
